Organize more the project and put together the extension in a single file.
